Refactored the project code
